modify project to new sql
