+++ PLMP-MVC/Views/User/Dashboard.cshtml
@@ -5,7 +5,6 @@
 
     var availableUnits = ViewBag.AvailableUnits as List<Unit> ?? new List<Unit>();
     var myRequests = ViewBag.MyRequests as List<MaintenanceRequest> ?? new List<MaintenanceRequest>();
-    var myApplications = ViewBag.MyApplications as List<Application> ?? new List<Application>();
     var myLeases = ViewBag.MyLeases as List<Lease> ?? new List<Lease>();
     var myPayments = ViewBag.MyPayments as List<Payment> ?? new List<Payment>();
 }
@@ -34,11 +33,11 @@
         display: inline-block;
     }
 
-        .btn-main:hover {
-            background: #1e40af;
-            color: white;
-            text-decoration: none;
-        }
+    .btn-main:hover {
+        background: #1e40af;
+        color: white;
+        text-decoration: none;
+    }
 
     .btn-pay {
         background: #f59e0b;
@@ -48,9 +47,9 @@
         border: none;
     }
 
-        .btn-pay:hover {
-            background: #e27d00;
-        }
+    .btn-pay:hover {
+        background: #e27d00;
+    }
 
     .section-header {
         margin-bottom: 20px;
@@ -119,10 +118,7 @@
                             <td>@unit.Type</td>
                             <td>@unit.RentAmount</td>
                             <td>
-                                <a asp-controller="User"
-                                   asp-action="Apply"
-                                   asp-route-id="@unit.UnitId"
-                                   class="btn-main">
+                                <a asp-controller="User" asp-action="Apply" asp-route-id="@unit.UnitId" class="btn-main">
                                     Apply
                                 </a>
                             </td>
@@ -157,30 +153,6 @@
     </div>
 
     <div class="card-box">
-        <h3 class="section-header">My Applications</h3>
-        @if (!myApplications.Any())
-        {
-            <p>No applications found.</p>
-        }
-        else
-        {
-            <table class="table">
-                <tbody>
-                    @foreach (var app in myApplications)
-                    {
-                        <tr>
-                            <td>@app.ApplicationId</td>
-                            <td>@app.UnitId</td>
-                            <td>@app.ApplicationDate</td>
-                            <td>@app.ApplicationStatus</td>
-                        </tr>
-                    }
-                </tbody>
-            </table>
-        }
-    </div>
-
-    <div class="card-box">
         <h3 class="section-header">My Leases</h3>
         @if (!myLeases.Any())
         {
@@ -190,12 +162,21 @@
         {
             <table class="table">
                 <tbody>
+                    <tr>
+                        <td>Lease ID</td>
+                        <td>Unit ID</td>
+                        <td>Lease Status</td>
+                        <td>Application Status</td>
+                    </tr>
+                </tbody>
+                <tbody>
                     @foreach (var lease in myLeases)
                     {
                         <tr>
                             <td>@lease.LeaseId</td>
                             <td>@lease.UnitId</td>
                             <td>@lease.LeaseStatus</td>
+                            <td>@lease.ApplicationStatus</td>
                         </tr>
                     }
                 </tbody>
